Bumped up almost all of the warnings in the TestApp project and fixed the resulting issues.

+++ FsprgEmbeddedStore/FsprgEmbeddedStoreController.m
@@ -209,7 +209,7 @@
 		return;
 	}
 	
-	float windowHeight = [[self webView] frame].size.height;
+	CGFloat windowHeight = [[self webView] frame].size.height;
 	id result = [[[self webView] windowScriptObject] evaluateWebScript:@"document.getElementsByClassName('store-page-navigation')[0].clientHeight"];
 	if (result == [WebUndefined undefined]) {
 		return;
@@ -220,7 +220,7 @@
 	float paddingTop = [[[cssStyle paddingBottom] substringToIndex:[[cssStyle paddingTop] length]-2] floatValue];
 	float paddingBottom = [[[cssStyle paddingBottom] substringToIndex:[[cssStyle paddingBottom] length]-2] floatValue];
 	
-	float newHeight = windowHeight - paddingTop - paddingBottom - pageNavigationHeight;
+	CGFloat newHeight = windowHeight - paddingTop - paddingBottom - pageNavigationHeight;
 	[[resizableContentE style] setHeight:[NSString stringWithFormat:@"%fpx", newHeight]];
 }
